Modular Refont UI

+++ CubeTools UI/Views/Actions/DeletePopUp.axaml.cs
@@ -1,14 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using Avalonia.Controls;
 using Avalonia.Input;
 using Avalonia.Interactivity;
 using Avalonia.Markup.Xaml;
 using CubeTools_UI.Models;
-using CubeTools_UI.Views.PopUps;
 using Library.ManagerExceptions;
-using Library.ManagerReader;
 using Library.ManagerWriter;
 using Library.Pointers;
 
@@ -16,8 +13,8 @@ namespace CubeTools_UI.Views.Actions
 {
     public class DeletePopUp : Window
     {
-        private MainWindowModel? _model;
-        private FileType _pointer;
+        private readonly LocalModel? _model;
+        private readonly FileType _pointer;
 
         #region Init
         
@@ -27,16 +24,13 @@ namespace CubeTools_UI.Views.Actions
             _model = null;
             _pointer = FileType.NullPointer;
         }
-        public DeletePopUp(MainWindowModel vm, FileType pointer) : this()
+        public DeletePopUp(LocalModel vm, FileType pointer) : this()
         {
             _model = vm;
             _pointer = pointer;
         }
 
-        private void InitializeComponent()
-        {
-            AvaloniaXamlLoader.Load(this);
-        }
+        private void InitializeComponent() => AvaloniaXamlLoader.Load(this);
         
         #endregion
 
@@ -67,23 +61,20 @@ namespace CubeTools_UI.Views.Actions
         #endregion
 
         
-
         private void DeletePointer()
         {
             // Create a new task to delete the pointer
             var task = new Task(() =>
             {
-                if (_pointer.IsDir)
-                    ManagerWriter.DeleteDir(_pointer);
-                else 
-                    ManagerWriter.Delete(_pointer);
+                if (_pointer.IsDir) ManagerWriter.DeleteDir(_pointer);
+                else ManagerWriter.Delete(_pointer);
             });
             // Remove reference from Directory Pointer
             _model?.ModelNavigationBar.DirectoryPointer.Remove(_pointer);
             // Run Tasks Async
             try
             {
-                if (_pointer.Size > 1000000)
+                if (_pointer.Size > 1000000 || _pointer.IsDir)
                 {
                     // Run async task
                     task.Start();
@@ -105,7 +96,7 @@ namespace CubeTools_UI.Views.Actions
                 if (exception is ManagerException @managerException)
                 {
                     @managerException.Errorstd = $"Unable to delete {_pointer.Name}";
-                    _model.View.SelectErrorPopUp(@managerException);
+                    _model?.SelectErrorPopUp(@managerException);
                 }
             }
         }
